+++ malachite-nz/src/test_util/gaussian_integer/arithmetic/pow.rs
@@ -6,10 +6,14 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-pub mod add;
-pub mod div_exact;
-pub mod div_rem;
-pub mod gcd;
-pub mod mul;
-pub mod pow;
-pub mod square;
+use crate::gaussian_integer::GaussianInteger;
+use malachite_base::num::basic::traits::One;
+
+// Repeated multiplication.
+pub fn gaussian_integer_pow_naive(x: &GaussianInteger, exp: u64) -> GaussianInteger {
+    let mut power = GaussianInteger::ONE;
+    for _ in 0..exp {
+        power *= x;
+    }
+    power
+}